+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **What Was Done:**
 
 - HATEOAS improvements
+- add POST to related documents
 
 
 ## [2.1.1](https://github.com/ivanoff/create-rest-api/tree/2.1.1) (2017-04-04)